% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_spec.R
\name{comp_spec}
\alias{comp_spec}
\title{Create a comparison specification}
\usage{
comp_spec(data)
}
\arguments{
\item{data}{A data frame (tibble recommended).}
}
\value{
An object of class \code{comp_spec}.
}
\description{
Create a comparison specification
}
\examples{
spec <- comp_spec(mtcars)
}
