% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compare_engines}
\alias{compare_engines}
\title{Compare multiple engines on the same spec}
\usage{
compare_engines(spec, engines = c("welch_t", "student_t", "mann_whitney"))
}
\arguments{
\item{spec}{A \code{comp_spec}.}

\item{engines}{Character vector of engine names.}
}
\description{
Compare multiple engines on the same spec
}
