% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.has_effectsize}
\alias{.has_effectsize}
\title{Check if the \code{effectsize} package is installed}
\usage{
.has_effectsize()
}
\value{
A logical scalar:
\itemize{
\item \code{TRUE} if the \pkg{effectsize} package is installed and loadable.
\item \code{FALSE} if it is not installed.
}
}
\description{
This internal helper is used to determine whether the \pkg{effectsize}
package is available in the current R session. It wraps
\code{\link[base:ns-load]{base::requireNamespace()}} with \code{quietly = TRUE} and returns \code{TRUE}
if the package can be loaded, \code{FALSE} otherwise.
}
\details{
This wrapper is useful for testing purposes as well.
}
\examples{
tidycomp:::.has_effectsize()

}
\keyword{internal}
