% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{.ph_dunn}
\alias{.ph_dunn}
\title{Dunn's test post-hoc comparisons (internal)}
\usage{
.ph_dunn(data, outcome, group, p.adjust, conf_level)
}
\description{
Requires the \code{rstatix} package.
}
\keyword{internal}
