% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{.ph_games_howell}
\alias{.ph_games_howell}
\title{Games-Howell post-hoc comparisons (internal)}
\usage{
.ph_games_howell(data, outcome, group, conf_level)
}
\description{
Requires the \code{rstatix} package.
}
\keyword{internal}
