% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{.ph_pairwise_t}
\alias{.ph_pairwise_t}
\title{Pairwise t-tests (internal)}
\usage{
.ph_pairwise_t(data, outcome, group, design, p.adjust, conf_level)
}
\description{
Used when \code{method = "pairwise_t_test"} to compute pairwise \emph{t}-tests with
the specified p-value adjustment.
}
\keyword{internal}
