% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{engine_mann_whitney}
\alias{engine_mann_whitney}
\title{Mann–Whitney (Wilcoxon rank-sum) engine}
\usage{
engine_mann_whitney(data, meta)
}
\description{
Mann–Whitney (Wilcoxon rank-sum) engine
}
\keyword{internal}
