% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_spec.R
\name{set_design}
\alias{set_design}
\title{Set design}
\usage{
set_design(spec, design = c("independent", "paired", "repeated", "factorial"))
}
\arguments{
\item{design}{One of "independent", "paired", "repeated", "factorial".}
}
\description{
Set design
}
