% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_spec.R
\name{set_strategy}
\alias{set_strategy}
\title{Set strategy (engine selection policy)}
\usage{
set_strategy(
  spec,
  strategy = c("auto", "pragmatic", "parametric", "robust", "permutation")
)
}
\arguments{
\item{spec}{A \code{comp_spec}.}

\item{strategy}{One of "auto", "pragmatic", "parametric", "robust", "permutation".}
}
\value{
The updated \code{comp_spec}.
}
\description{
Choose how the engine will be selected when running \code{\link[=test]{test()}}.
}
\examples{
spec <- comp_spec(mtcars)
spec <- set_strategy(spec, "auto")
}
