% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{tidycomp-package}
\alias{tidycomp}
\alias{tidycomp-package}
\title{tidycomp: Tidyverse-Style Functions for Statistical Comparisons and Diagnostics}
\description{
Tidy, auditable pipelines for group comparisons. Tidycomp streamlines two-sample, paired, one-way, and repeated-measures designs, delivering standardised results and effect sizes. With flexible outlier trimming and clear, friendly messaging, it makes diagnostics and inference transparent, composable, and seamless within a tidymodels-like ecosystem.
}
\author{
\strong{Maintainer}: Daniel Dauber \email{daniel.dauber.io@gmail.com} (\href{https://orcid.org/0000-0003-4324-4796}{ORCID}) [copyright holder]

}
\keyword{internal}
