% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_spec.R
\name{set_design}
\alias{set_design}
\title{Set design}
\usage{
set_design(spec, design = c("independent", "paired", "repeated", "factorial"))
}
\arguments{
\item{spec}{A \code{comp_spec}.}

\item{design}{One of "independent", "paired", "repeated", "factorial".}
}
\value{
The updated \code{comp_spec}.
}
\description{
Define the comparison design on a \code{comp_spec}.
}
\examples{
spec <- comp_spec(mtcars)
spec <- set_design(spec, "independent")
}
