% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_spec.R
\name{set_outcome_type}
\alias{set_outcome_type}
\title{Set outcome type}
\usage{
set_outcome_type(spec, type = c("numeric", "binary", "ordered", "count"))
}
\arguments{
\item{type}{One of "numeric", "binary", "ordered", "count".}
}
\description{
Set outcome type
}
