% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_spec.R
\name{set_strategy}
\alias{set_strategy}
\title{Set strategy (engine selection policy)}
\usage{
set_strategy(
  spec,
  strategy = c("auto", "pragmatic", "parametric", "robust", "permutation")
)
}
\arguments{
\item{strategy}{One of "auto","pragmatic","parametric","robust","permutation".}
}
\description{
Set strategy (engine selection policy)
}
